only self-proxy cy.visits matching a net-stubbing route
works around over-application of `selfProxy` opt
