+++ packages/net-stubbing/lib/server/index.ts
@@ -8,6 +8,8 @@ export { InterceptResponse } from './intercept-response'
 
 export { NetStubbingState } from './types'
 
+export { getRouteForRequest } from './route-matching'
+
 import { state } from './state'
 
 export const netStubbingState = state
